% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_dist_heatmap.R
\name{cat_dist_heatmap}
\alias{cat_dist_heatmap}
\title{Create concatenated charts showing the heatmap of two categorical variables and the barcharts for occurrance of these variables.
Heatmap will be on the left and the two barcharts will be on the right in the same column.}
\usage{
cat_dist_heatmap(
  cat_1,
  cat_2,
  data,
  title = "",
  lab_1 = "",
  lab_2 = "",
  heatmap = TRUE,
  barchart = TRUE
)
}
\arguments{
\item{cat_1}{Name of the column name for the first categorical variable.}

\item{cat_2}{Name of the column name for the second categorical variable.}

\item{data}{Target data frame for visualization.}

\item{title}{Title for the chart.}

\item{lab_1}{Axis label for the first categorical variable (x-axis).}

\item{lab_2}{Axis label for the second categorical variable (y-axis).}

\item{heatmap}{Whether to include a heatmap plot or not.}

\item{barchart}{Whether to include the barchart or not.}
}
\value{
A concatenated ggplot chart consists of a heatmap and 2 barcharts.
@examples
cat_dist_heatmap(
cat_1 = school_type, cat_2 = program_type, data = data,
title = 'School Type vs Program Type',
lab_1 = 'School type', lab_2 = 'Program type', heatmap = True, barchart = True)
}
\description{
Create concatenated charts showing the heatmap of two categorical variables and the barcharts for occurrance of these variables.
Heatmap will be on the left and the two barcharts will be on the right in the same column.
}
