% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_dist_by_cat.R
\name{num_dist_by_cat}
\alias{num_dist_by_cat}
\title{Create a histogram paired with a box plot.}
\usage{
num_dist_by_cat(
  num,
  group,
  data,
  title_hist = "",
  title_boxplot = "",
  lab_num = "",
  lab_group = "",
  num_on_x = TRUE,
  stat = TRUE
)
}
\arguments{
\item{num}{Name of the column for the numeric variable.}

\item{group}{Name of the column for the categorical variable.}

\item{data}{Target data frame for visualization.}

\item{title_hist}{Title for the histogram.}

\item{title_boxplot}{Title for the boxplot.}

\item{lab_num}{Axis label for the numeric variable.}

\item{lab_group}{Axis label for the categorical variable.}

\item{num_on_x}{Whether the numeric variable is put on the x-axis in the boxplot.}

\item{stat}{Whether printing the test statistic and summary or not.}
}
\value{
A ggplot object consists of a histogram and a boxplot.
}
\description{
Create a histogram paired with a box plot.
}
\examples{
data <- data.frame( num = c( 1, 2, 3, 4), classes = c( 1, 1, 2, 2))
num_dist_by_cat(
  num = num, group = classes, data = data,
  title_hist = 'Distribution of X', title_boxplot = 'X Seperated by Group',
  lab_num = 'X', lab_group = 'Group', num_on_x = TRUE, stat = TRUE)
}
