% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_dist_summary.R
\name{num_dist_summary}
\alias{num_dist_summary}
\title{Create a distribution plot of the numeric variable
in general and statistical summary  of the feature .
In addition, the  correlation values of the input
variable with other features based on a threshold will also be returned}
\usage{
num_dist_summary(
  col_name,
  data,
  title = "",
  label_x = "",
  thresh_corr = 0,
  stat = TRUE
)
}
\arguments{
\item{col_name}{A string  with the name of the column}

\item{data}{A dataframe , target DataSet for visualization}

\item{title}{A string , Title for chart , default =''}

\item{label_x}{A string,  X-axis label for numeric column}

\item{thresh_corr}{Number Threshold value for correlation}

\item{stat}{Boolean, whether to show statistic information or not}
}
\value{
plot and dataframe to display summary statistics and correlation values
@examples
num_dist_summary(
col_name =num1, data=df, title='num_col1',label_x= 'distribution of num1', thresh_corr = 0.3,stat = TRUE
)
}
\description{
Create a distribution plot of the numeric variable
in general and statistical summary  of the feature .
In addition, the  correlation values of the input
variable with other features based on a threshold will also be returned
}
