% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_dist_by_cat.R
\name{num_dist_by_cat}
\alias{num_dist_by_cat}
\title{Create a histogram paired with a box plot.}
\usage{
num_dist_by_cat(
  v_num,
  v_cat,
  data,
  title_hist = "",
  title_boxplot = "",
  lab_num = "",
  lab_cat = "",
  num_on_x = TRUE,
  stat = TRUE
)
}
\arguments{
\item{v_num}{Name of the column for the numeric variable.}

\item{v_cat}{Name of the column for the categorical variable.}

\item{data}{Target data frame for visualization.}

\item{title_hist}{Title for the histogram.}

\item{title_boxplot}{Title for the boxplot.}

\item{lab_num}{Axis label for the numeric variable.}

\item{lab_cat}{Axis label for the categorical variable.}

\item{num_on_x}{Whether the numeric variable is put on the x-axis in the boxplot.}

\item{stat}{Whether printing the test statistic and summary or not.}
}
\value{
A ggplot object consists of a histogram and a boxplot.
}
\description{
Create a histogram paired with a box plot.
}
\examples{
num_dist_by_cat(
  v_num = num, v_cat = groups, data = data,
  title_hist = 'Distribution of X', title_boxplot = 'X Seperated by Group',
  lab_num = 'X', lab_cat = 'Group', num_on_x = True, stat = True)
}
