% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_dist_scatter.R
\name{num_dist_scatter}
\alias{num_dist_scatter}
\title{num_dist_scatter
Creates a scatter plot given two numerical features.}
\usage{
num_dist_scatter(
  num1,
  num2,
  data,
  title = "",
  stat = FALSE,
  trend = FALSE
)
}
\arguments{
\item{num1}{Name of the column for the first numeric feature.}

\item{num2}{Name of the column for the second numeric feature.}

\item{data}{Target data frame for visualization.}

\item{title}{Title for the chart.}


\item{stat}{Whether or not to include simple statistics.}

\item{trend}{Type of trend line. Options are: FALSE, 'lm', 'loess'.}
}
\value{
A ggplot scatter plot object
}
\description{
num_dist_scatter
Creates a scatter plot given two numerical features.
}
\examples{
num_dist_scatter(
  num1=num1, num2=num1, data=df, title='num_col1 v num_col2', stat = TRUE, trend='lin'
)
}
